requirements.txt
selenium-4.8.0
requests-2.28.1